[MOV] purchase_stock: create stock_move own file
Separates the inheritance of the `stock.move` model in `purchase_stock`
in its own file.
