+++ addons/purchase_stock/models/__init__.py
@@ -8,4 +8,5 @@ from . import res_config_settings
 from . import res_partner
 from . import res_company
 from . import stock
+from . import stock_move
 from . import stock_rule
